+++ api/openapi_server/util.py
@@ -1,7 +1,16 @@
+import os
+import ntpath
 import datetime
 
 import six
 import typing
+
+
+def join_read_path_with_data_dir(read_path):
+    """
+    Joins a read path with the data directory
+    """
+    return os.path.join(os.getenv('DATA_DIR', os.getcwd()), ntpath.basename(read_path))
 
 
 def _deserialize(data, klass):